// src/types/member.ts
export interface Member {
  id: number;
  CardCode: string;
  full_name: string;
  GroupCode: number;
  CardType: string;
  phone: string;
  phone2: string;
  email: string;
  id_number: number;
  kra_pin: string;
  isRegFeePaid: number;
  date_of_joining: string;
  subscription_balance: number;
  total_savings: number;
  loan_balance: number;
  opening_balance: number;
  date_of_closing: string | null;
  gender: string;
  nationality: string;
  dob: string;
  marital_status: string;
  ocupation: string;
  address: string;
  address2: string;
  apartment_name: string;
  postal_address: string;
  employer_name: string;
  employer_address: string;
  employer_phone: string;
  employer_phone_2: string;
  position: string;
  employment_type: string;
  monthly_income: number;
  others_income: number;
  monthly_expenses: number;
  nok_full_name: string;
  nok_relation: string;
  nok_phone: string;
  nok_email: string;
  nok_address: string;
  mode_of_contribution: string;
  payment_mode: string | null;
  bank_name: string;
  bank_branch: string;
  bank_account_number: string;
  status: number;
  created_at: string;
  updated_at: string;
  documents: any[];
}